requirements.txt
Django>=5.0,<6.0
Pillow>=10.0
django-storages>=1.14
boto3>=1.34
